Moved tests to test suite
