+++ Cargo.toml
@@ -52,6 +52,7 @@ ironrdp-testsuite-core = { path = "crates/ironrdp-testsuite-core" }
 ironrdp-tls = { version = "0.1", path = "crates/ironrdp-tls" }
 ironrdp-tokio = { version = "0.1", path = "crates/ironrdp-tokio" }
 ironrdp = { version = "0.5", path = "crates/ironrdp" }
+now-proto-pdu = { version = "0.1", path = "crates/now-proto-pdu" }
 
 bitflags = "2.4"
 expect-test = "1"
