+++ crates/ironrdp-testsuite-core/src/now_proto.rs
@@ -1,17 +1,15 @@
 //! Various test utilities
-use alloc::vec::Vec;
 use expect_test::Expect;
 
 use ironrdp_pdu::cursor::ReadCursor;
 use ironrdp_pdu::PduDecode as _;
 
-use crate::NowMessage;
+use now_proto_pdu::NowMessage;
 
-pub(crate) fn now_msg_roundtrip(msg: impl Into<NowMessage>, expected_bytes: Expect) {
+pub fn now_msg_roundtrip(msg: impl Into<NowMessage>, expected_bytes: Expect) {
     let msg = msg.into();
 
-    let mut buf = Vec::new();
-    let _ = ironrdp_pdu::encode_buf(&msg, &mut buf).unwrap();
+    let buf = ironrdp_pdu::encode_vec(&msg).unwrap();
 
     expected_bytes.assert_eq(&format!("{:02X?}", buf));
