+++ crates/ironrdp-testsuite-core/tests/now_proto/mod.rs
@@ -1,0 +1,6 @@
+mod exec;
+mod number;
+mod session;
+mod status;
+mod string;
+mod system;
