+++ crates/ironrdp-testsuite-core/tests/pcb.rs
@@ -99,8 +99,7 @@ fn null_size() {
         .err()
         .unwrap();
 
-    expect![
-        [r#"
+    expect![[r#"
             Error {
                 context: "PreconnectionBlob",
                 kind: InvalidMessage {
@@ -109,8 +108,7 @@ fn null_size() {
                 },
                 source: None,
             }
-        "#]
-    ]
+        "#]]
     .assert_debug_eq(&e);
 }
