+++ crates/now-proto-pdu/src/core/string.rs
@@ -296,75 +296,3 @@ pub type NowString128 = NowRestrictedStr<127>;
 ///
 /// NOW-PROTO: NOW_STRING256
 pub type NowString256 = NowRestrictedStr<255>;
-
-#[cfg(all(test, feature = "std"))]
-mod tests {
-    use super::*;
-    use rstest::rstest;
-
-    #[rstest]
-    #[case("hello", &[0x05, 0x00, 0x00, 0x00, b'h', b'e', b'l', b'l', b'o', 0x00])]
-    #[case("", &[0x00, 0x00, 0x00, 0x00, 0x00])]
-    fn now_lrgstr_roundtrip(#[case] value: &str, #[case] expected_encoded: &'static [u8]) {
-        let mut encoded_value = [0u8; 32];
-        let encoded_size = crate::encode(&NowLrgStr::new(value).unwrap(), &mut encoded_value).unwrap();
-
-        assert_eq!(encoded_size, expected_encoded.len());
-        assert_eq!(&encoded_value[..encoded_size], expected_encoded);
-
-        let decoded_value = crate::decode::<NowLrgStr>(&encoded_value).unwrap();
-        assert_eq!(decoded_value.0, value);
-    }
-
-    #[test]
-    fn decoded_now_lrgstr_invalid_utf8() {
-        let encoded = [0x01, 0x00, 0x00, 0x00, 0xFF, 0x00];
-        crate::decode::<NowLrgStr>(&encoded).unwrap_err();
-    }
-
-    #[rstest]
-    #[case("hello", &[0x05, b'h', b'e', b'l', b'l', b'o', 0x00])]
-    #[case("", &[0x00, 0x00])]
-    fn now_varstr_roundtrip(#[case] value: &str, #[case] expected_encoded: &'static [u8]) {
-        let mut encoded_value = [0u8; 32];
-        let encoded_size = crate::encode(&NowVarStr::new(value).unwrap(), &mut encoded_value).unwrap();
-
-        assert_eq!(encoded_size, expected_encoded.len());
-        assert_eq!(&encoded_value[..encoded_size], expected_encoded);
-
-        let decoded_value = crate::decode::<NowVarStr>(&encoded_value).unwrap();
-        assert_eq!(decoded_value.0, value);
-    }
-
-    #[test]
-    fn decoded_now_varstr_invalid_utf8() {
-        let encoded = [0x01, 0xFF, 0x00];
-        crate::decode::<NowLrgStr>(&encoded).unwrap_err();
-    }
-
-    #[rstest]
-    #[case("hello", &[0x05, b'h', b'e', b'l', b'l', b'o', 0x00])]
-    #[case("", &[0x00, 0x00])]
-    fn now_restricted_string_roundtrip(#[case] value: &str, #[case] expected_encoded: &'static [u8]) {
-        let mut encoded_value = [0u8; 32];
-        let encoded_size = crate::encode(&NowString16::new(value).unwrap(), &mut encoded_value).unwrap();
-
-        assert_eq!(encoded_size, expected_encoded.len());
-        assert_eq!(&encoded_value[..encoded_size], expected_encoded);
-
-        let decoded_value = crate::decode::<NowString16>(&encoded_value).unwrap();
-        assert_eq!(decoded_value.0, value);
-    }
-
-    #[test]
-    fn now_restricted_string_too_large_constructed() {
-        let value = "a".repeat(16);
-        NowString16::new(value).unwrap_err();
-    }
-
-    #[test]
-    fn decoded_now_restricted_string_invalid_utf8() {
-        let encoded = [0x01, 0xFF, 0x00];
-        crate::decode::<NowString16>(&encoded).unwrap_err();
-    }
-}
